+++ packages/docs/src/stories/Avatar.stories.tsx
@@ -5,7 +5,10 @@ export default {
   title: 'Data display/Avatar',
   component: Avatar,
 
-  args: {},
+  args: {
+    src: 'https://github.com/pugialli.png',
+    alt: 'João Paulo Pugialli',
+  },
 
   argTypes: {
     src: {
@@ -18,18 +21,29 @@ export default {
         type: 'text',
       },
     },
+    variant: {
+      options: ['sm', 'md', 'lg'],
+      control: {
+        type: 'inline-radio',
+      },
+    },
   },
 } as Meta<AvatarProps>
 
 export const Primary: StoryObj<AvatarProps> = {
   args: {
-    src: 'https://github.com/pugialli.png',
-    alt: 'João Paulo Pugialli',
+    variant: 'md',
   },
 }
 
+export const Big: StoryObj<AvatarProps> = {
+  args: {
+    variant: 'lg',
+  },
+}
 export const WithFallback: StoryObj<AvatarProps> = {
   args: {
     src: undefined,
+    alt: undefined,
   },
 }