+++ packages/docs/src/stories/Checkbox.stories.tsx
@@ -11,7 +11,7 @@ export default {
       return (
         <Box
           as="label"
-          css={{ display: 'flex', flexDirection: 'row', gap: '$2' }}
+          css={{ display: 'flex', flexDirection: 'row', gap: '$2', width:'90%', alignItems:'center' }}
         >
           {Story()}
           <Text size="sm">Accept terms of use</Text>