+++ packages/docs/src/stories/Text.stories.tsx
@@ -13,19 +13,15 @@ export default {
   argTypes: {
     size: {
       options: [
-        'xxs',
         'xs',
         'sm',
         'md',
         'lg',
         'xl',
         '2xl',
+        '3xl',
         '4xl',
         '5xl',
-        '6xl',
-        '7xl',
-        '8xl',
-        '9xl',
       ],
       control: {
         type: 'inline-radio',