+++ packages/react/src/components/Avatar/index.tsx
@@ -1,12 +1,14 @@
 import { ComponentProps } from 'react'
 import { AvatarContainer, AvatarFallback, AvatarImage } from './styles'
-import { User } from 'phosphor-react'
+import { User } from "@phosphor-icons/react"
 
-export interface AvatarProps extends ComponentProps<typeof AvatarImage> {}
+export interface AvatarProps extends ComponentProps<typeof AvatarImage> {
+  variant: 'sm' | 'md' | 'lg'
+}
 
-export function Avatar(props: AvatarProps) {
+export function Avatar({variant, ...props}: AvatarProps) {
   return (
-    <AvatarContainer>
+    <AvatarContainer variant={variant}>
       <AvatarImage {...props} />
 
       <AvatarFallback delayMs={600}>