+++ packages/react/src/components/Avatar/styles.ts
@@ -4,9 +4,28 @@ import { styled } from '../../styles'
 export const AvatarContainer = styled(Avatar.Root, {
   borderRadius: '$full',
   display: 'inline-block',
-  width: '$16',
-  height: '$16',
   overflow: 'hidden',
+
+  variants: {
+    variant: {
+      sm: {
+        width: '$6',
+        height: '$6',
+      },
+      md: {
+        width: '$12',
+        height: '$12',
+      },
+      lg: {
+        width: '$40',
+        height: '$40',
+      },
+    },
+  },
+  
+  defaultVariants: {
+    variant: 'md',
+  },
 })
 
 export const AvatarImage = styled(Avatar.Image, {