+++ packages/react/src/components/Heading.tsx
@@ -5,18 +5,15 @@ export const Heading = styled('h2', {
   fontFamily: '$default',
   lineHeight: '$shorter',
   margin: 0,
-  color: '$gray100',
+  color: '$gray900',
 
   variants: {
     size: {
-      sm: { fontSize: '$xl' },
-      md: { fontSize: '$2xl' },
-      lg: { fontSize: '$4xl' },
-      '2xl': { fontSize: '$5xl' },
-      '3xl': { fontSize: '$6xl' },
-      '4xl': { fontSize: '$7xl' },
-      '5xl': { fontSize: '$8xl' },
-      '6xl': { fontSize: '$9xl' },
+      sm: { fontSize: '$lg' },
+      md: { fontSize: '$xl' },
+      lg: { fontSize: '$3xl' },
+      '2xl': { fontSize: '$4xl' },
+      '3xl': { fontSize: '$5xl' },
     },
   },