+++ packages/react/src/components/TextArea.tsx
@@ -2,22 +2,22 @@ import { ComponentProps } from 'react'
 import { styled } from '../styles'
 
 export const TextArea = styled('textarea', {
-  backgroundColor: '$gray900',
+  backgroundColor: '$gray100',
   padding: '$3 $4',
   borderRadius: '$sm',
   boxSizing: 'border-box',
-  border: '2px solid $gray900',
+  border: '2px solid $gray500',
 
   fontFamily: '$default',
   fontSize: '$sm',
-  color: '$white',
+  color: '$gray700',
   fontWeight: '$regular',
   resize: 'vertical',
   minHeight: 80,
 
   '&:focus': {
     outline: 0,
-    borderColor: '$ignite300',
+    borderColor: '$secondary',
   },
 
   '&:disabled': {