+++ packages/tokens/src/font-sizes.ts
@@ -4,10 +4,8 @@ export const fontSizes = {
   md: '1rem', //16px
   lg: '1.125rem', //18px
   xl: '1.25rem', //20px
-  xxl: '1.313rem', //21px - remover
   '2xl': '1.5rem', //24px
-  '3xl': '1.625rem', //26px - remover
-  '4xl': '2rem', //32px
-  '5xl': '2.25rem', //36px
-  '6xl': '3rem', //48px
+  '3xl': '2rem', //32px
+  '4xl': '2.25rem', //36px
+  '5xl': '3rem', //48px
 }