+++ packages/react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ejc-ui/react
 
+## 1.0.0
+
+### Major Changes
+
+- Tokens updated and React Components Redone
+
 ## 0.2.1
 
 ### Patch Changes