+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ejc-ui/tokens
 
+## 1.1.0
+
+### Minor Changes
+
+- Tokens updated and React Components Redone
+
 ## 1.0.0
 
 ### Major Changes